requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
jinja2>=3.1.0
python-multipart>=0.0.9
aiosmtplib>=3.0.0
pillow>=10.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
ruff>=0.1.0
